Add: an upload API to AzurStat for CN users
