+++ module/config/argument/args.json
@@ -100,6 +100,14 @@
         "type": "input",
         "value": null
       },
+      "API": {
+        "type": "select",
+        "value": "normal",
+        "option": [
+          "normal",
+          "cn_reverse_proxy"
+        ]
+      },
       "ResearchRecord": {
         "type": "select",
         "value": "do_not",
